--- Helper/HttpClientFactory.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Net.Http;
-
-namespace promotion.Helper
-{
-    public class HttpClientFactory : IHttpClientFactory
-    {
-        public HttpClient Create()
-        {
-            return new HttpClient
-            {
-                BaseAddress = new Uri(Constants.ReqresEndpoint)
-            };
-        }
-    }
-}
