--- Helper/IHttpClientFactory.cs
@@ -1,9 +1,0 @@
-﻿using System.Net.Http;
-
-namespace promotion.Helper
-{
-    public interface IHttpClientFactory
-    {
-        HttpClient Create();
-    }
-}
