--- Helper/UriPaths.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace promotion.Helper
-{
-    public static class UriPaths
-    {
-        public const string Rest = "/api/rest";
-        public const string GetUserEndpoint = "/api/users/{0}";
-    }
-}
