+++ Library/UriConfiguration.cs
@@ -11,6 +11,7 @@
         }
 
         public string Host { get; set; }
+
         public string Path { get; set; } = "/";
     }
 }
