+++ Models/PostData.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace promotion.Models
+﻿namespace promotion.Models
 {
     public class PostData
     {
