+++ ProxyHttp/UriPaths.cs
@@ -1,0 +1,7 @@
+﻿namespace promotion.ProxyHttp
+{
+    public static class UriPaths
+    {
+        public const string GetUserEndpoint = "/api/users/{0}";
+    }
+}
